requirements.txt
numpy
pandas
pyyaml
scikit-learn
matplotlib
tensorflow
tqdm
pytest
lightgbm